refactor(model): keep stop policy in reflection
